tests/wpt/web-platform-tests/XMLHttpRequest/access-control-basic-allow-star.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Tests "*" setting for Access-Control-Allow-Origin header</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script type="text/javascript">
    const xhr = new XMLHttpRequest;

    test(function(test) {
      xhr.open("GET", get_host_info().HTTP_REMOTE_ORIGIN + "/XMLHttpRequest/resources/access-control-basic-allow-star.py", false);

      xhr.send();

      assert_equals(xhr.responseText, "PASS: Cross-domain access allowed.");
    }, "Allow star");
    </script>
  </body>
</html>
